# fix: Fix perf issue with command log