--- a/README.md
+++ b/README.md
@@ -1,4 +1,3 @@
-#
 <p align="center">
   <a href="https://www.cypress.io"><img src="https://cloud.githubusercontent.com/assets/1268976/20607953/d7ae489c-b24a-11e6-9cc4-91c6c74c5e88.png"/></a>
 </p>