--- a/cli/schema/cypress.schema.json
+++ b/cli/schema/cypress.schema.json
@@ -30,6 +30,11 @@
           "default": 50,
           "description": "The number of tests for which snapshots and command data are kept in memory. Reduce this number if you are experiencing high memory consumption in your browser during a test run."
         },
+        "numCommandsToShow": {
+          "type": "number",
+          "default": null,
+          "description": "The number of tests to show in the command log. Reduce this number if you are experiencing slowness with long running tests"
+        },
         "port": {
           "type": "number",
           "default": null,