--- a/packages/reporter/src/hooks/hooks.tsx
+++ b/packages/reporter/src/hooks/hooks.tsx
@@ -56,9 +56,11 @@ export interface HookProps {
 }
 
 const Hook = observer(({ model, showNumber }: HookProps) => {
+  // If a user specified `numCommandsToShow` in the cypress config
+  // then only show the last `numCommandsToShow` in the command log
+  // This would primarily be done as a performance optimization
   const numCommandsToShow = parseInt(Cypress.config('numCommandsToShow'), 10)
-  const parsedNumCommandsToShow = isNaN(numCommandsToShow) ? undefined : numCommandsToShow;
-  const commandsToShow = parsedNumCommandsToShow ? model.commands.slice(numCommandsToShow * -1) : model.commands;
+  const commandsToShow = numCommandsToShow ? model.commands.slice(numCommandsToShow * -1) : model.commands;
   return <li className={cs('hook-item', { 'hook-failed': model.failed, 'hook-studio': model.isStudio })}>
     <Collapsible
       header={<HookHeader model={model} number={showNumber ? model.hookNumber : undefined} />}