… packages/config/__snapshots__/index_spec.js
… packages/server/test/unit/config_spec.js