<!-- Thanks for contributing! PLEASE...
- Read our contributing guidelines: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md 
- Read our Code Review Checklist on coding standards and what needs to be done before a PR can be merged: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md#Code-Review-Checklist
- Mark this PR as "Draft" if it is not ready for review.
- Make sure you set the correct base branch based on what packages you're changing: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md#branches
-->

- Closes (I tagged a bunch of performance issues that I think this will address) <!-- link to the issue here, if there is one -->
    - https://github.com/cypress-io/cypress/issues/16130
    - https://github.com/cypress-io/cypress/issues/8658
    - https://github.com/cypress-io/cypress/issues/6270
    - https://github.com/cypress-io/cypress/issues/6388
    - https://github.com/cypress-io/cypress/issues/15880
    - https://github.com/cypress-io/cypress/issues/5987
### User facing changelog
<!-- 
Explain the change(s) for every user to read in our changelog. Examples: https://on.cypress.io/changelog
If the change is not user-facing, write "n/a".
-->

Added a new configuration option, `numCommandsToShow`, which can be used to yield performance improvements for long running tests.

This will limit the number to commands that show in the Command Log (the left sidebar) of a cypress test, instead of allowing it to grow indefinitely.


### Additional details
<!-- Examples:
- Why was this change necessary?
- What is affected by this change?
- Any implementation details to explain?
-->

We already have a parameter, `CYPRESS_NO_COMMAND_LOG`, which hides the Command Log altogether. However, this is not sufficient for both debugging and performance.  `numCommandsToShow` seeks to provide a solution that satisfies both performance and debugging concerns.

While this does restrict the user from clicking on snapshots past the specified `numCommandsToShow` value (and only if they opted into this), I don't think this is typically needed (when debugging, we only need the last couple of commands around when the test failed)

Related docs PR:
https://github.com/cypress-io/cypress-documentation/pull/4431

### How has the user experience changed?
<!-- Provide before and after examples of the change.
Screenshots or GIFs are preferred. -->

I believe this simple change / configuration can yield huge huge performance gains for users, with little to no risk of regression at all.

I ran a cypress test w/o this change, and repeated the same series of steps in a test 1000 times (I just took screenshots because the video was too big, and I also didn't bother waiting to get to 1000).

The test was to:
* Go to a public todo page
* Add a todo item
* Delete the todo item
* Repeat

From the screenshots, we can see that this step started off taking 400ms, and very quickly we found ourselves at the 1200ms mark, a 3x increment from when the test started, and after only 100 iterations.

#### Before applying this PR
<img width="1344" alt="Screen Shot 2022-04-08 at 6 09 32 PM" src="https://user-images.githubusercontent.com/3589158/162539330-413d990d-c13d-469f-bd34-973e9d745993.png">

<img width="1344" alt="Screen Shot 2022-04-08 at 6 10 44 PM" src="https://user-images.githubusercontent.com/3589158/162539361-2ce20b61-7222-4305-8d0a-1cdbb222dae3.png">

#### After applying this PR (using a `numCommandsToShow` of 25)

Next, I applied this PR, and adjusted the new configuration option, numCommandsToShow,  to 25.

For the duration of the test, the duration of the same sequence of actions did not linearly increase with the number of commands getting run.  Also, note that at the 100 iteration mark, this test took /50%/ of the time as the previous test, 42s vs 87s.

<img width="1344" alt="Screen Shot 2022-04-08 at 6 13 36 PM" src="https://user-images.githubusercontent.com/3589158/162540120-df1dca00-b1c4-4c3a-b0cd-86d100257d71.png">

<img width="1344" alt="Screen Shot 2022-04-08 at 6 14 07 PM" src="https://user-images.githubusercontent.com/3589158/162540137-913f832a-3656-42bf-8616-9e01819b3d7d.png">

### PR Tasks
<!-- 
These tasks must be completed before a PR is merged.
If a task does not apply, write [na] instead of checking the box.
DO NOT DELETE the PR checklist.
-->

- [x] Have tests been added/updated?
- [ ] Has the original issue (or this PR, if no issue exists) been tagged with a release in ZenHub? (user-facing changes only)
- [x] Has a PR for user-facing changes been opened in [`cypress-documentation`](https://github.com/cypress-io/cypress-documentation)? <!-- Link to PR here -->
- [x] Have API changes been updated in the [`type definitions`](https://github.com/cypress-io/cypress/blob/develop/cli/types/cypress.d.ts)?
- [x] Have new configuration options been added to the [`cypress.schema.json`](https://github.com/cypress-io/cypress/blob/develop/cli/schema/cypress.schema.json)?