@andrewmtam 2022-04-11
Thanks @jennifer-shehane , appreciate the feedback!  I've got two follow up questions if you've got a moment:

1. As we already have a `CYPRESS_NO_COMMAND_LOG={1|0}` environment variable which does something similar, how do you feel about adding a `CYPRESS_LIMIT_COMMAND_LOG={number}` environment variable (instead of it being a configuration option). I believe this environment variable could provide more benefits over the one that hides the command log altogether, as it still surfaces information surrounding failures, /and/ it speeds up performance.  I know this just shifts your argument about:

> We would prefer a solution that addresses the performance issues without degrading the experience of using the entirety of the Cypress Test Runner at the time it is run 

out of a configuration option and into an environment variable, but seeing that we already have an environment variable that hides the command log altogether, I just wanted to double check whether or not doing the same thing for limiting the command log is also off the table.  My main push for this point is that, if consumers set `CYPRESS_NO_COMMAND_LOG` and find that there is a significant performance boost, there is not really a next step for them (while it fixes their perf issues, it cannot be run in production because it does not surface any error messaging), and so providing an alternate `CYPRESS_LIMIT_COMMAND_LOG` may be a way to bridge that gap ([some requests here regarding that](https://github.com/cypress-io/cypress/issues/6270)). 

2. If that doesn't sound good, another alternative could be to look into virtualization of some sort (since we know the command log can grow indefinitely, it sounds like we'd want to curb that, as opposed to continuing to render everything, but faster).  If that's an approach you are aligned with, I'd be happy to poke around there next time I have a few cycles