@andrewmtam 2022-04-11
Ooo one more idea came to mind:

3. By default, limit the command log to show X items (lets say 25), but surface "Show more" and/or "Show all" button at the top of the command list for that section that a user can interact with if they need to see more commands. This sounds like it could be a good happy medium:
* It provides a nice performance boost out of the box
* If users need to see more commands, they can opt-in by clicking "Show more" within the same instance of the test run

Something kinda like this:
<img width="1081" alt="Screen Shot 2022-04-11 at 11 25 11 AM" src="https://user-images.githubusercontent.com/3589158/162773349-0cf00b3d-ed8b-4bb0-81bf-5f01a0765fe8.png">

Clicking "Show More" each time could increase the amount shown by 25. So on first click, 50 commands would show, then 75, 100, etc etc.  But if the user really does want to see everything, they just click "Show All", and it'll behave as-if th command log was never limited