@andrewmtam 2022-04-12
Hey @jennifer-shehane , I wanted to give one more go at this and put up a PR here:

https://github.com/cypress-io/cypress/pull/21036

This does not implement the ideal solution you hinted at, but it does solve this problem:
> Not being able to see x commands degrades the experience of the runner significantly as information is essentially lost from that run.

by providing a way to surface the entire command log list and allow users to click through snapshots from the whole test run. Curious to hear your thoughts on this approach if you've got a couple minutes!